+++ core/src/main/kotlin/com/dllewellyn/coinbaseapi/models/currency/OpenOrder.kt
@@ -3,20 +3,20 @@ package com.dllewellyn.coinbaseapi.models.currency
 import com.dllewellyn.coinbaseapi.models.BuyOrSell
 import com.dllewellyn.coinbaseapi.models.trade.LimitOrMarket
 import com.dllewellyn.coinbaseapi.models.trade.OrderStatus
-import java.math.BigInteger
+import java.math.BigDecimal
 
 data class OpenOrder(
     val createdAt: String,
-    val executedValue: BigInteger,
-    val fillFees: BigInteger,
-    val fillSize: BigInteger,
+    val executedValue: BigDecimal,
+    val fillFees: BigDecimal,
+    val fillSize: BigDecimal,
     val id: String,
     val postOnly: Boolean,
-    val price: BigInteger,
+    val price: BigDecimal,
     val currencyPair: CurrencyPair,
     val settled: Boolean,
     val side: BuyOrSell,
-    val size: BigInteger,
+    val size: BigDecimal,
     val status: OrderStatus,
     val stp: String?,
     val timeInForce: String,
