+++ core/src/main/kotlin/com/dllewellyn/coinbaseapi/models/trade/CurrencyBuyAndSell.kt
@@ -1,5 +1,5 @@
 package com.dllewellyn.coinbaseapi.models.trade
 
-import java.math.BigInteger
+import java.math.BigDecimal
 
-data class CurrencyBuyAndSell(val currencyFrom : String, val currencyTo : String, val buy : BigInteger, val sell : BigInteger)
+data class CurrencyBuyAndSell(val currencyFrom : String, val currencyTo : String, val buy : BigDecimal, val sell : BigDecimal)
